+++ src/main/java/com/coupang/marketplace/cart/repository/CartRepository.java
@@ -1,6 +1,5 @@
 package com.coupang.marketplace.cart.repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.apache.ibatis.annotations.Mapper;
@@ -14,7 +13,7 @@ public interface CartRepository {
 
 	void insertProduct(Cart cart);
 
-	void updateProductNum(BigInteger productId, int productNum);
+	void updateProductNum(long productId, int productNum);
 
-	Optional<Cart> findByProductId(BigInteger userId, BigInteger productId);
+	Optional<Cart> findByProductId(long userId, long productId);
 }
